+++ CHANGELOG.md
@@ -350,3 +350,4 @@
 - [Issue #386](https://github.com/influxdb/influxdb/issues/386). Drop series should work with series containing dots
 - [Issue #389](https://github.com/influxdb/influxdb/issues/389). Filtering shouldn't stop prematurely
 - [Issue #341](https://github.com/influxdb/influxdb/issues/341). Make the number of shards that are queried in parallel configurable
+- [Issue #394](https://github.com/influxdb/influxdb/issues/394). Support count(distinct) and count(DISTINCT)